+++ lib/js_of_ocaml/dom_html.ml
@@ -414,6 +414,13 @@ and touch =
     method pageY : int readonly_prop
   end
 
+and submitEvent =
+  object
+    inherit event
+
+    method submitter : element t optdef readonly_prop
+  end
+
 and dragEvent =
   object
     inherit mouseEvent
@@ -1011,7 +1018,7 @@ class type formElement =
 
     method reset : unit meth
 
-    method onsubmit : ('self t, event t) event_listener writeonly_prop
+    method onsubmit : ('self t, submitEvent t) event_listener writeonly_prop
   end
 
 class type optGroupElement =